Swapped login and policy accept page
By doing the policy accept dialog after the login, we can do that
together with the page that we need to set the login cookie.

Note: we can't set the login cookie directly in the response the user
was bounced to by the OAuth 2.0 server, as in strict settings a browser
won't allow to do this (e.g. Firefox in privacy mode).